feat: 스크롤 시 효과 뜨도록 수정
related to: #160
