+++ src/Onboarding/components/frame/Frame2.tsx
@@ -1,27 +1,32 @@
 import { css } from '@emotion/react';
 
 import imgFrame2TreeGraph from '../../assets/imgFrame2TreeGraph.png';
+import useScrollDetect from '../../hooks/useScrollDetect';
 import { ImgPopUp } from '../../styles/animation';
 import { sectionStyle } from '../../styles/common';
 import TextField from './TextField';
 
 const Frame2 = () => {
+  const { active, element } = useScrollDetect();
+
   return (
-    <section css={section}>
+    <section css={section} ref={element}>
       <TextField
         subTitle="OKR tree"
         subTitleColor="main_purple"
         title={`흩어져 있는 할 일을\n하나의 북극성 아래`}
         description="파편화된 할일들을 단 3개의 우선순위로 정렬시켜 핵심 지표를 달성하는데 집중하세요"
       />
-      <ImgPopUp
-        fromY={20}
-        toY={0}
-        src={imgFrame2TreeGraph}
-        alt="tree-img"
-        width={986}
-        height={657}
-      />
+      {active && (
+        <ImgPopUp
+          fromY={20}
+          toY={0}
+          src={imgFrame2TreeGraph}
+          alt="tree-img"
+          width={986}
+          height={657}
+        />
+      )}
     </section>
   );
 };
