+++ src/Onboarding/components/frame/Frame5.tsx
@@ -1,4 +1,4 @@
-import { css } from '@emotion/react';
+import { css, keyframes } from '@emotion/react';
 import styled from '@emotion/styled';
 import { useEffect, useState } from 'react';
 
@@ -8,12 +8,14 @@ import imgFrame5HistoryTask from '../../assets/imgFrame5HistoryTask.png';
 import imgFrame5List2 from '../../assets/imgFrame5List2.png';
 import imgFrame5List3 from '../../assets/imgFrame5List3.png';
 import imgFrame5List4 from '../../assets/imgFrame5List4.png';
+import useScrollDetect from '../../hooks/useScrollDetect';
 import { ImgPopUp } from '../../styles/animation';
 import { sectionStyle } from '../../styles/common';
 import TextField from './TextField';
 
 const Frame5 = () => {
   const [isOpen, setIsOpen] = useState(false);
+  const { active, element } = useScrollDetect();
 
   useEffect(() => {
     const timer = setTimeout(() => {
@@ -24,57 +26,59 @@ const Frame5 = () => {
   }, []);
 
   return (
-    <section css={section}>
+    <section css={section} ref={element}>
       <TextField
         subTitle="History"
         subTitleColor="sub_blue"
         title="나의 모든 우주 여정을 여기에"
         description="꾸준함으로 쌓인 발자취를 확인하며 앞으로의 미래를 향해 다시 나아가세요"
       />
-      <div css={imgContainer}>
-        {isOpen && (
-          <img src={imgFrame5HistoryOpen} alt="main-dashboard-img-1" width={1005} height={57} />
-        )}
-        {!isOpen && (
-          <img src={imgFrame5HistoryClose} alt="main-dashboard-img-1" width={1005} height={57} />
-        )}
-        <LateImgPopUp
-          fromY={5.7}
-          toY={7.2}
-          delay={2.2}
-          src={imgFrame5HistoryTask}
-          alt="main-dashboard-img-2"
-          width={1005}
-          height={168}
-        />
-        <ImgPopUp
-          fromY={0}
-          toY={19}
-          delay={2}
-          src={imgFrame5List2}
-          alt="main-dashboard-img-3"
-          width={1005}
-          height={57}
-        />
-        <ImgPopUp
-          fromY={0}
-          toY={19}
-          delay={2}
-          src={imgFrame5List3}
-          alt="main-dashboard-img-3"
-          width={1005}
-          height={57}
-        />
-        <ImgPopUp
-          fromY={0}
-          toY={19}
-          delay={2}
-          src={imgFrame5List4}
-          alt="main-dashboard-img-3"
-          width={1005}
-          height={57}
-        />
-      </div>
+      {active && (
+        <ImgContainer>
+          {isOpen && (
+            <img src={imgFrame5HistoryOpen} alt="main-dashboard-img-1" width={1005} height={57} />
+          )}
+          {!isOpen && (
+            <img src={imgFrame5HistoryClose} alt="main-dashboard-img-1" width={1005} height={57} />
+          )}
+          <LateImgPopUp
+            fromY={5.7}
+            toY={7.2}
+            delay={2.2}
+            src={imgFrame5HistoryTask}
+            alt="main-dashboard-img-2"
+            width={1005}
+            height={168}
+          />
+          <ImgPopUp
+            fromY={0}
+            toY={19}
+            delay={2}
+            src={imgFrame5List2}
+            alt="main-dashboard-img-3"
+            width={1005}
+            height={57}
+          />
+          <ImgPopUp
+            fromY={0}
+            toY={19}
+            delay={2}
+            src={imgFrame5List3}
+            alt="main-dashboard-img-3"
+            width={1005}
+            height={57}
+          />
+          <ImgPopUp
+            fromY={0}
+            toY={19}
+            delay={2}
+            src={imgFrame5List4}
+            alt="main-dashboard-img-3"
+            width={1005}
+            height={57}
+          />
+        </ImgContainer>
+      )}
     </section>
   );
 };
@@ -90,7 +94,16 @@ const section = css`
   ${sectionStyle}
 `;
 
-const imgContainer = css`
+const fadeIn = keyframes`
+    from {
+      opacity: 0;
+    }
+    to {
+      opacity: 1;
+    }
+  `;
+
+const ImgContainer = styled.div`
   position: relative;
   display: flex;
   flex-direction: column;
@@ -98,6 +111,7 @@ const imgContainer = css`
   align-items: center;
   justify-content: center;
   width: 100vw;
+  animation: ${fadeIn} 1s;
 `;
 
 const LateImgPopUp = styled(ImgPopUp)`
